requirements.txt
tensorflow==2.0.0
Pillow==6.2.0
